Refactoring of main Offers page
