+++ components/OfferSubFilter.tsx
@@ -1,0 +1,158 @@
+import { FilterWithCount, QuestionFilter } from "../lib/model/FilterModel";
+import cx from "classnames";
+
+const styles = {
+	defaultButton:
+		"text-gray-900 font-medium py-1 px-2 border rounded-3xl m-1 text-sm bg-white border-gray-200 hover:bg-blue-50",
+	activeButton:
+		"bg-blue-600 text-white border-blue-800 shadow-sm hover:bg-blue-600",
+};
+
+const OfferSubFilter: React.FC<{
+	shownFilters: Array<FilterWithCount>;
+	questionFilter: QuestionFilter;
+	setQuestionFilter: React.Dispatch<React.SetStateAction<QuestionFilter>>;
+}> = ({ shownFilters, questionFilter, setQuestionFilter }) => {
+	return (
+		<>
+			{shownFilters.map((filter) => {
+				const selectedGroups = filter.optionGroups?.filter((group) =>
+					group.options.some((option) =>
+						(questionFilter[filter.id] ?? []).includes(option)
+					)
+				);
+				const selectedGroupsOptions = selectedGroups?.flatMap(
+					(it) => it.options
+				);
+				const shownOptions =
+					selectedGroupsOptions !== undefined
+						? filter.options.filter((it) =>
+								selectedGroupsOptions.includes(it.id)
+						  )
+						: filter.options;
+
+				return (
+					<div key={filter.id} className="mt-4">
+						<h3 className="text-center">{filter.question}</h3>
+						{filter.optionGroups && (
+							<ul className="mt-1 flex flex-wrap justify-center">
+								<li>
+									<button
+										className={cx(
+											styles.defaultButton,
+											!(questionFilter[filter.id] ?? []).length &&
+												styles.activeButton
+										)}
+										onClick={() => {
+											setQuestionFilter((state) => ({
+												...state,
+												[filter.id]: [],
+											}));
+										}}
+									>
+										Nezáleží
+									</button>
+								</li>
+
+								{filter.optionGroups.map((option) => {
+									const selected = selectedGroups!.includes(option);
+									return (
+										<li key={option.id}>
+											<button
+												className={cx(
+													styles.defaultButton,
+													selected && styles.activeButton
+												)}
+												onClick={() => {
+													if (selected) {
+														setQuestionFilter((state) => ({
+															...state,
+															[filter.id]: state[filter.id]!.filter(
+																(it) => !option.options.includes(it)
+															),
+														}));
+													} else {
+														setQuestionFilter((state) => ({
+															...state,
+															[filter.id]: [
+																...(state[filter.id] ?? []),
+																...option.options,
+															],
+														}));
+													}
+												}}
+											>
+												{option.label} ({option.count})
+											</button>
+										</li>
+									);
+								})}
+							</ul>
+						)}
+
+						{shownOptions.length > 0 && (
+							<ul className="mt-1 flex flex-wrap justify-center">
+								{!filter.optionGroups && (
+									<li>
+										<button
+											className={cx(
+												styles.defaultButton,
+												!(questionFilter[filter.id] ?? []).length &&
+													styles.activeButton
+											)}
+											onClick={() => {
+												setQuestionFilter((state) => ({
+													...state,
+													[filter.id]: [],
+												}));
+											}}
+										>
+											Nezáleží
+										</button>
+									</li>
+								)}
+								{shownOptions.map((option) => {
+									const selected = (questionFilter[filter.id] ?? []).includes(
+										option.id
+									);
+									return (
+										<li key={option.id}>
+											<button
+												className={cx(
+													styles.defaultButton,
+													selected && styles.activeButton
+												)}
+												onClick={() => {
+													if (selected) {
+														setQuestionFilter((state) => ({
+															...state,
+															[filter.id]: state[filter.id]!.filter(
+																(it) => it !== option.id
+															),
+														}));
+													} else {
+														setQuestionFilter((state) => ({
+															...state,
+															[filter.id]: [
+																...(state[filter.id] ?? []),
+																option.id,
+															],
+														}));
+													}
+												}}
+											>
+												{option.label} ({option.count})
+											</button>
+										</li>
+									);
+								})}
+							</ul>
+						)}
+					</div>
+				);
+			})}
+		</>
+	);
+};
+
+export default OfferSubFilter;
