+++ package.json
@@ -32,6 +32,7 @@
 		"@types/node": "17.0.21",
 		"@types/react": "17.0.39",
 		"autoprefixer": "^10.4.4",
+		"classnames": "^2.3.1",
 		"cypress": "^9.5.2",
 		"eslint": "8.10.0",
 		"eslint-config-next": "12.1.0",
